introduce ThemeManager, which should always be used for fallback styles. Also moving classes around
