+++ lib/layout/box_layout.dart
@@ -1,15 +1,18 @@
+import 'package:ensemble/framework/action.dart';
 import 'package:ensemble/framework/widget/widget.dart';
-import 'package:ensemble/layout/layout_helper.dart';
 import 'package:ensemble/layout/templated.dart';
 import 'package:ensemble/page_model.dart';
 import 'package:ensemble/screen_controller.dart';
 import 'package:ensemble/util/layout_utils.dart';
 import 'package:ensemble/util/utils.dart';
+import 'package:ensemble/widget/helpers/theme_manager.dart';
 import 'package:ensemble_ts_interpreter/invokables/invokable.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart' as flutter;
 import 'package:ensemble/util/platform.dart';
+
+import '../widget/helpers/controllers.dart';
 
 class Column extends BoxLayout {
   static const type = 'Column';
@@ -93,6 +96,55 @@ abstract class BoxLayout extends StatefulWidget with UpdatableContainer, Invokab
 
   bool isVertical();
 
+}
+
+/// base controller for Column/Row
+class BoxLayoutController extends BoxController {
+  EnsembleAction? onTap;
+
+  bool scrollable = false;
+  bool autoFit = false;
+  String? mainAxis;
+  String? crossAxis;
+  String? mainAxisSize;
+  int? maxWidth;
+  int? maxHeight;
+  int? gap;
+
+  String? fontFamily;
+  int? fontSize;
+
+  List<Widget>? children;
+
+  // applicable to Flex container only
+  String? direction;
+  // applicable only for ListView
+  EnsembleAction? onItemTap;
+  Color? sepratorColor;
+  double? sepratorWidth;
+  EdgeInsets? sepratorPadding;
+  int selectedItemIndex= -1;
+
+
+  @override
+  Map<String, Function> getBaseSetters() {
+    Map<String, Function> setters = super.getBaseSetters();
+    setters.addAll({
+      'scrollable': (value) => scrollable = Utils.getBool(value, fallback: false),
+      'autoFit': (value) =>  autoFit = Utils.getBool(value, fallback: false),
+      'mainAxis': (value) => mainAxis = Utils.optionalString(value),
+      'crossAxis': (value) => crossAxis = Utils.optionalString(value),
+      'mainAxisSize': (value) => mainAxisSize = Utils.optionalString(value),
+      'maxWidth': (value) => maxWidth = Utils.optionalInt(value),
+      'maxHeight': (value) => maxHeight = Utils.optionalInt(value),
+      'gap': (value) => gap = Utils.optionalInt(value),
+
+      'fontFamily': (value) => fontFamily = Utils.optionalString(value),
+      'fontSize': (value) =>  fontSize = Utils.optionalInt(value),
+
+    });
+    return setters;
+  }
 }
 
 class BoxLayoutState extends WidgetState<BoxLayout> with TemplatedWidgetState {
@@ -270,7 +322,7 @@ class BoxLayoutState extends WidgetState<BoxLayout> with TemplatedWidgetState {
         borderRadius: widget._controller.borderRadius != null ? widget._controller.borderRadius!.getValue() : null,
         boxShadow: widget._controller.shadowColor == null ? null : <BoxShadow>[
           BoxShadow(
-            color: Color(widget._controller.shadowColor!),
+            color: widget._controller.shadowColor ?? ThemeManager.getShadowColor(context),
             blurRadius: (widget._controller.shadowRadius ?? 0).toDouble(),
             offset: widget._controller.shadowOffset ?? const Offset(0, 0),
           )
